Use double not float

+++ src/no/runsafe/moosic/MusicHandler.java
@@ -30,7 +30,7 @@ public class MusicHandler
 		TrackPlayer trackPlayer = new TrackPlayer(location, musicTrack, volume);
 		this.currentTrackPlayerID++;
 
-		long delay = (long) ((float) (speed / musicTrack.getTempo()) * 20);
+		long delay = (long) ((double) (speed / musicTrack.getTempo()) * 20);
 		Plugin.output.write("Delay: " + delay);
 		int timer = this.scheduler.startSyncRepeatingTask(new Runnable() {
 			@Override
